requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

httpx == 0.22.0
loguru == 0.6.0
m3u8 == 1.0.0
pycryptodomex == 3.14.1
typer == 0.4.0
